Rework DashboardSplashApp component.
